Renamed 'main-sql-temp.py' to 'main.py'
